intermediate result
