+++ lib/domain/model/advanced_image.dart
@@ -1,0 +1,7 @@
+final class AdvancedImage {
+  final String url;
+  final String id;
+  final DateTime? createdAt;
+
+  const AdvancedImage({required this.url, required this.id, this.createdAt});
+}
