+++ lib/main.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:surf_flutter_summer_school_24/feature/theme/data/theme_repository.dart';
-import 'package:surf_flutter_summer_school_24/feature/theme/domain/theme_controller.dart';
-import 'package:surf_flutter_summer_school_24/feature/theme/di/theme_inherited.dart';
-import 'package:surf_flutter_summer_school_24/screens/image_carousel_screen.dart';
-import 'package:surf_flutter_summer_school_24/screens/main_screen.dart';
+import 'package:surf_flutter_summer_school_24/data/mock_image_repository.dart';
+import 'package:surf_flutter_summer_school_24/data/theme_repository.dart';
+import 'package:surf_flutter_summer_school_24/domain/interactors/advanced_image_interactor.dart';
+import 'package:surf_flutter_summer_school_24/domain/model/theme_controller.dart';
+import 'package:surf_flutter_summer_school_24/di/theme_inherited.dart';
+import 'package:surf_flutter_summer_school_24/ui/image_carousel_screen.dart';
+import 'package:surf_flutter_summer_school_24/ui/main_screen.dart';
 import 'package:surf_flutter_summer_school_24/storage/theme/theme_storage.dart';
-import 'package:surf_flutter_summer_school_24/feature/theme/ui/theme_builder.dart';
+import 'package:surf_flutter_summer_school_24/ui/theme_builder.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -22,6 +24,7 @@ void main() async {
       child: const MyApp(),
     ),
   );
+
 }
 
 class MyApp extends StatelessWidget {
