+++ pubspec.yaml
@@ -7,9 +7,7 @@ environment:
   sdk: '>=3.4.3 <4.0.0'
 
 dependencies:
-  carousel_slider: ^4.2.1
   collection: ^1.18.0
-  cupertino_icons: ^1.0.8
   flutter:
     sdk: flutter
   shared_preferences: ^2.2.3
